[FIX] BO get all financing

+++ src/main/java/com/danamon/autochain/entity/Tenure.java
@@ -27,7 +27,7 @@ public class Tenure {
     private FinancingPayable financingPayableId;
 
     @Column
-    private Long Amount;
+    private Double Amount;
 
     @Column
     private Date dueDate;
